Add a symmetry breaker
This imposes a canonical order on sets to prevent
symmetric solutions. This might help speed up
solves for sets of things where order doesn’t
matter, e.g. transitions.

This uses a technique pertained to in this paper:
https://www.cs.cmu.edu/~wklieber/papers/2007_efficient-cnf-encoding-for-selecting-1.pdf

And elaborated here:
http://cstheory.stackexchange.com/questions/17139/how-do-i-use-canonical-ordering-to-reduce-symmetry-in-the-sat-encoding-of-the-pi

However, this might not actually help Lingeling as
it is already clever enough to deal with trivial
symmetries, see:
http://fmv.jku.at/papers/Biere-SAT-Competition-2013-TPH-problem.pdf

Whilst benchmarking I found it helped Riss but
hindered Lingeling for the trivial test case. I’ll
have to test with/without symmetry breaking when
it comes to run the thing.