Remove the constraint about left/right symbols
This was interfering with another constraint in
the connection reducer. In fact, we don’t need 
this constraint because the connection reducer now
ensures the head cannot go off the left or right
edge of the tape, so we don’t care what value the
left/right symbols take at positions 0 and n-1,
respectively.